Convert preferences.js to typescript +html updates
